requirements.txt
flask
pymongo
tabulate
Pillow
opencv-python
requests
matplotlib
pandas